2004-10-21	blytkerchan	compile vector impl with stricter compiler 	* vector.h [vector_t]: change types of size_t elements to uint32_t 	* vector.c: prefer local copies to installed ones for libmemory 	headers 	(internal_resize): reflect changes to vector_t
move list node functions to their own files
	* list_node.c: new file
	* list_node.h: ditto
	* list.c: include list_node.h
	[list_state_t]: move to list_node.h
	(list_node_free): move to list_node.c
	(list_node_new): ditto
	(list_node_insert): ditto
	(list_find): ditto, but rename to list_node_find
	(list_insert): update calls to list_node_insert
	(list_delete): call list_node_find in stead of list_find
	(list_search): ditto
	* list.h: include list_node.h
	[list_node_t]: move to list_node.h
	* module.mk [libcontain_SRC]: add list_node.c

+++ vector.c
@@ -33,8 +33,8 @@
  */
 #include <stdlib.h>
 #include <string.h>
-#include <libmemory/smr.h>
-#include <libmemory/hptr.h>
+#include "libmemory/smr.h"
+#include "libmemory/hptr.h"
 #include "arch/include/set.h"
 #include "arch/include/compare_and_exchange.h"
 #include "arch/include/increment.h"
@@ -43,7 +43,7 @@
 #include "binary.h"
 #include "thread.h"
 
-static void internal_resize(vector_t * vector, size_t n_size);
+static void internal_resize(vector_t * vector, uint32_t n_size);
 static void internal_condense(vector_t * vector);
 
 /* This may seem obvious, and it is! Allocate a new vector of SIZE entries */
@@ -218,12 +218,12 @@ size_t vector_get_numentries(vector_t * vector)
 	return retval;
 }
 
-static void internal_resize(vector_t * vector, size_t n_size)
+static void internal_resize(vector_t * vector, uint32_t n_size)
 {
 	vector_node_t * new_nodes = calloc(n_size, sizeof(vector_node_t));
 	vector_node_t * o_nodes;
 	vector_node_t * nodes;
-	size_t size;
+	uint32_t size;
 	
 	do
 	{
